Add -c option to trigger S-LDM when a vehicle is outside internal area
This commit introduces a new option (-c), following the discussion
the CLC partners had on Friday April 23rd 2021.
When this option is specified, a special "cross-border trigger
mode" is activated, enabling the triggering of the S-LDM also
due to vehicles which are not located inside its internal coverage
area (i.e. vehicles located inside the extended part of the coverage
area).
This may help the cross-border UC handling in case a vehicle cannot
subscribe to multiple AMQP brokers (in different countries) at once.