Add a first implementation of a REST client for the Maneuvering Service
This commit introduces a first implementation of a possible REST client
for the maneuvering service (for a Cooperative Lane Change use case),
together with the first "Trigger Manager", i.e. a module checking if
some triggering condition is occurring.
The module introduced here is a simple "Trigger Manager", which
considers as "triggering condition" the activation of a turn indicator.
If a triggering condition occurs, this module manages the creation
of a proper REST client to send the needed data to a Maneuvering
Service, acting as REST server.
The triggering module is disabled by default, for the time being.