DENM example in main.cpp and updated GeoNet in decoder-module
This commits adds an example of DENM in the "Test area" of main.cpp.
This is just a sample DENM meant to further evaluate the decoder and
the related code will be removed from the final version of the S-LDM.
This commit also introduces a small modification to the GeoNetworking
module in the decoder-module. Now all the packets with the protocol
version set to "0" will just generate a warning instead of being
discarded (even though the expected version value should always
be "1", according to the ETSI standards).