Improve context generation around vehicle of interest
This commit improves the context generation for the Maneuvering
Service, which is now vehicle-centric (on the triggering vehicle),
instead of being based on a fixed central point.
It also improves how the "turnindicator" information is sent via
the REST API: if no indicator is on, "off" is sent instead of
sending an empty string ("").