Update the checkDim() function in QuadKeyTS.cpp
This commit updates the checkDim() function in QuadKeyTS to fix
some bugs which could cause a wrong computation of the "compressed
Quadkeys" (in case some kind of compression was needed to reduce
the overall length of the filter string, to be sent to the AMQP
broker).
This fix has been done by Lorenzo De Persiis.