Add options for settings ports and IP addresses and fix triggering
This commits adds several new command line options to make the
S-LDM completely configurable through the command line and avoid
using any hardcoded IP address or port number.
This commit also fixes the triggering mechanism by checking
whether a triggering vehicle is located inside the internal area
covered by the current instance of the S-LDM.
If it is not (i.e., if it is located in the external area), no
triggering is performed, unless the very specific "cross_border_trigger"
option is specified (this option is disabled by default).
The mechanism of triggering only when the vehicle is located
inside the internal S-LDM area has been agreed as part of the
5G-CARMEN service architecture.