Implement vehicle visualizer object removal
This commit implements a way to remove old objects from the
vehicle visualizer map (i.e., old objects which are automatically
removed from the database by the DBcleaner_callback() thread).
The modifications include the addition of a new function inside
the DB module (LDMMap): deleteOlderThanAndExecute().
This commit also updates the Dockerfile to download and build
Qpid Proton 0.35.0 (instead of 0.33.0).