Create a second implementation of unifyQuadkeys
This commit adds a second implementation of unifyQuadkeys, starting
from all the logic of the first implementation.
This second implementation (unifyQuadkeys2()) has been created as an
attempt to quickly fix a bug which was found in the main implementation
during some road tests.
This commit also slightly modifies the m_latlon_variation values, to
ensure that Quadkeys are never skipped when "scanning" the values
of latitude and longitude.