Enable subscription to multiple AMQP brokers
This commit enables the subscription to multiple AMQP brokers, with the aim
of tackling cross-border situations, in which an S-LDM instance may need
to subscribe to both a local AMQP broker and to a broker hosted by other
operators belonging to neighboring contries, in order to have a complete
view of the road over the border.
This commit also introduces few additional options and a way to check
if the data being added to the database is up-to-date or not (based
on the GNTimestamp values received from vehicles). This check can
be optionally disabled with the --disable-age-check/-g option.
This commit also updates the UDP->AMQP relayer to accept a new
command line option (--comm-port or -P) to specify the UDP port for
the communication with ms-van3t (instead of hardcoding the default
value, i.e. 20000).