Add AMQP idle timeout options
This commit adds few options (both for the main AMQP client and
for the possible additional ones) to manage the AMQP idle timeout.
These options allow the user to set an idle timeout and also to
attempt an automatic reconnection when local idle timeout errors
occurs inside Qpid Proton.
This commit also updates the Makefile to delete the Quadkey cache
file (cachefile.sldmc) when the "fullclean" target is invoked.