Upgrading the message decoder to support also "pure" Facilities Layer messages
This commit upgrades the message decoder to support not only the
decoding of full ITS messages (i.e., with BTP and GeoNetworking),
but also the one of "pure" Facilities Layer messages, i.e., CAMs
and DENMs which are directly encapsulated inside AMQP 1.0, without
BTP and GeoNetworking.
This commit also changes the S-LDM version to v1.1.2 and updates
the UDP-AMQP relayer, adding a new experimental option to test
the transmission of "pure" CAMs to the S-LDM, without BTP and
GeoNetworking ("--skip-gn"). This option is still experimental
and it works when only CAMs are transmitted from ms-van3t to the
UDP-AMQP relayer.