Manage the retrieval of exterior lights information from  1.3.2 version CAMs
This enables the S-LDM to extract the ext. lights information from older
versions of CAM for which the lowFreqContainer is not possible to decode.
This changes enables teh ability to extract this information from the
driveDirection field in the highSpeedContainer used with a custom encoding.
This functionality is enabled with an additional option.