Add implementation of on-demand JSON-over-TCP interface
This commit introduces a new interface for retrieving data
from the S-LDM on demand, without relying on the triggering
mechanisms.
When this interface is activated, the S-LDM exposes a server
waiting for TCP packets containing a proper JSON request.
Upon reception of a proper request, the S-LDM sends as reply
(i.e., as a JSON structure inside a TCP packet) a list of
vehicles located within a given geographical area.